handles permissions being rejected or revoced
app does not crach when revocing location permssion during runtime,
closes #77.
checks implemented to ensure that the google API client is not created
before location permissions are given, closes #79.
When starting the app it checks if googleAPI can be used, if not Patriks
version of location gathering is used, closes #81.
Formatted Mainactivity and structured the functions logically.