fixes null pictures when rorating screen
manifest updated so that a new activity does not start as soon as the
screen rotates when taking a picture. This ensures that the picture is
stored and converted to png as it should, closes #107.

Changed decodeSampledBitmapFromFile to return a uri instead of a bitmap,
this to clean up the mainactivity.

We should make main more threaded, now this error comes when null error
was before:
I/Choreographer: Skipped 288 frames!  The application may be doing too
much work on its main thread.