Thread for problem picture
Async implementation had worse performance than creating a new thread
for problem picture creation. When creating a thread the program
continues from the camera intent much faster than with a AsyncTask,
closes #120.

AsyncTask: up to 2500ms to create picture
Thread: up to 1500ms to create picture (this should not matter as the
user does not notice any lag in the mainthread.)

A side effect of this is that the up dateUrisHashmap needs to be static
for the other thread to be able to access it.

Another side effect is that gpsPicture needs to be public in order for
other thread to access it.

General code cleanup in main.