Duct-tape solution for app crash
Look at issue 127. I simply removed toggleSendButton call from
updateUrisHashmap and the app no longer crashes. However, this reopens
issues 31, as the sendButton will not update correctly. We'll have to
figure this one out at the hackaton/retrospect meeting, either we
disable the need for taking a picture for now, or find a correct
solution that doesn't crash the app.

I also removed redundant permission-asking of phone number and contacts
write, as we do not use them at all.

I also made the help toasts shorter and fixed the finnish language
version to be correct finnish.